chore: remove migration that would have required doctrine and add changes to initial migration instead

+++ database/migrations/2024_03_23_070019_add_quantities_to_recipe_items.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('recipe_item', function (Blueprint $table) {
-            $table->integer('quantity')->nullable()->default(1);
+            $table->float('quantity')->nullable()->default(1);
             $table->foreignId('quantity_unit_id')->nullable()->constrained('quantity_units');
         });
     }
